DRAG-01 organize components in folders
